Add old media deletion flag

+++ controllers/actions/SingleMediaAction.php
@@ -2,16 +2,22 @@
 
 namespace h3tech\crud\controllers\actions;
 
+use h3tech\crud\controllers\MediaController;
 use h3tech\crud\models\Media;
-use yii\base\InvalidConfigException;
+use Throwable;
+use yii\base\Exception;
 use yii\db\ActiveRecord;
+use yii\db\StaleObjectException;
 use yii\web\UploadedFile;
 
 class SingleMediaAction extends MediaAction
 {
+    /**
+     * @throws Exception
+     */
     protected function uploadMedia(ActiveRecord $model, UploadedFile $mediaFile)
     {
-        $controllerClass = $this->controllerClass;
+        /** @var MediaController $mediaControllerClass */
         $mediaControllerClass = $this->getMediaControllerClass();
 
         $model->{$this->mediaIdAttribute} = $mediaControllerClass::upload(
@@ -26,6 +32,9 @@ class SingleMediaAction extends MediaAction
         return UploadedFile::getInstance($model, $this->fileVariable);
     }
 
+    /**
+     * @throws Exception
+     */
     public function afterCreate(ActiveRecord $model)
     {
         $mediaFile = $this->getMediaFile($model);
@@ -34,24 +43,32 @@ class SingleMediaAction extends MediaAction
         }
     }
 
+    /**
+     * @throws Throwable
+     * @throws StaleObjectException
+     */
     public function afterUpdate(ActiveRecord $model)
     {
         $mediaFile = $this->getMediaFile($model);
         if ($mediaFile !== null) {
-            $oldMedia = Media::findOne($model->{$this->mediaIdAttribute});
+            $oldMediaId = $model->{$this->mediaIdAttribute};
 
             $this->uploadMedia($model, $mediaFile);
-            $model->save();
 
-            if ($oldMedia !== null) {
+            if ($this->deleteOldMedia && ($oldMedia = Media::findOne($oldMediaId)) !== null) {
+                $model->save();
                 $oldMedia->delete();
             }
         }
     }
 
+    /**
+     * @throws Throwable
+     * @throws StaleObjectException
+     */
     public function afterDelete(ActiveRecord $model)
     {
-        if (($media = Media::findOne($model->{$this->mediaIdAttribute})) !== null) {
+        if ($this->deleteOldMedia && ($media = Media::findOne($model->{$this->mediaIdAttribute})) !== null) {
             $media->delete();
         }
     }
